Update App.java
Use Java ImageIO instead of reading raw bmp bytes. Also rewrote the for loops to read the image pixel-by-pixel, and refactored the reading and writing of pixels to separate methods. An rgba image is converted to 1-bit color space by discarding the alpha channel, converting the rgb to monochrome with (r+g+b)/3 and setting the pixel if (r+g+b)/3 >= 127.